[ML] Rename analytics package to dataframe
