+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/action/TransportRunAnalyticsAction.java
@@ -38,9 +38,9 @@ import org.elasticsearch.transport.TransportService;
 import org.elasticsearch.xpack.core.ml.action.RunAnalyticsAction;
 import org.elasticsearch.xpack.core.ml.utils.ExceptionsHelper;
 import org.elasticsearch.xpack.ml.MachineLearning;
-import org.elasticsearch.xpack.ml.analytics.DataFrameDataExtractorFactory;
-import org.elasticsearch.xpack.ml.analytics.DataFrameFields;
-import org.elasticsearch.xpack.ml.analytics.process.AnalyticsProcessManager;
+import org.elasticsearch.xpack.ml.dataframe.DataFrameDataExtractorFactory;
+import org.elasticsearch.xpack.ml.dataframe.DataFrameFields;
+import org.elasticsearch.xpack.ml.dataframe.process.AnalyticsProcessManager;
 
 import java.util.Arrays;
 import java.util.Collections;
