+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/dataframe/DataFrameDataExtractorFactory.java
@@ -3,7 +3,7 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-package org.elasticsearch.xpack.ml.analytics;
+package org.elasticsearch.xpack.ml.dataframe;
 
 import org.elasticsearch.ResourceNotFoundException;
 import org.elasticsearch.action.ActionListener;
